feat: validate if default sort is valid
